Timing: add new `getDurationSince()` method
This method doesn't use the "system" run start time, but uses an arbitrary starttime as provided via a parameter.

This allows for removing some duplicate code in various places.

Includes making the class `final` and explicitly marking this class as an internal class to allow for further changes in the future.