Squiz/VariableComment: make errorcode for "TagNotAllowed" warning modular
As things are, the `VariableComment` sniff only allows `@see` tags and `@var` tags in property docblocks.

This practice feels dated to me, as commonly used tags, such as `@link`, `@deprecated` etc are not allowed.

This commit replaces the generic `TagNotAllowed` error code in the sniff with a more modular `[TagName]TagNotAllowed` error code.

This allows for rulesets to selectively ignore the `TagNotAllowed` error code for specific tags the ruleset maintainer wants to allow for the code base under scan.

Includes tests.