PSR2/SwitchDeclaration: update doc as `case` body should be non-empty
This commit updates the valid and invalid code examples in a
`<code_comparison>` block, as the `There must be a comment when
fall-through is intentional in a non-empty case body` error is only
triggered for non-empty `case` bodies. It also fixes the positioning of
the `<em>` tags in the modified code examples.