StandardRulesetsQATest: fix the test
With the change to the `StatusWriter` for writing error output to `STDERR`, this test was no longer working correctly as the output bufffering was only catching output send to `STDOUT`, which means the protection this test was supposed to give was no undone.

Fixed now. And this includes removing yet some more output buffering.