Config: bug fix - "reports" could get set incorrectly when in CBF mode
There are three CLI arguments related to the reporting:
* `--report=[comma separated list]`
* `--report-file=[path to file]`
* `--report-<type>=[path to file]`

The `--report-file...` option is ignored when in CBF mode. Whether that is the correct behaviour or not, is outside the scope of the current bug fix.

In the `Config::processLongArgument()` method, the `--report-file` option is handled _before_ the `--report-<type>` option to ensure that `--report-file` is handled correctly, as the handling of the `--report-<type>` argument is based on seeing the prefix `--report-`, which overlaps with `--report-file`.

However, this logic was broken in CBF mode, as in that case, the `--report-file` CLI arg would still fall through to the `--report-<type>` handling, which would break the setting of the `$reports`.

Fixed now.

Includes tests specifically for this issue.