GH Actions/verify release: tweak the workflow [2]
PHIVE does not support a "stability" flag yet, which means that it will always grab the very latest release, even when this is an alpha/beta/RC pre-release.

As the GH API `releases/latest` route returns the latest _stable_ release, this causes the workflow to fail.
This was already previously identified as an area which needed attention in https://github.com/PHPCSStandards/PHP_CodeSniffer/pull/758, and as we now have a pre-release as the latest release (`4.0.0beta1`), this now needs to be fixed.

Note: this still doesn't fix things completely, as in reality, the workflow should work based on the tag published if the workflow is triggered by a release to ensure that the "release assets" and "versioned web" jobs check the right release ("unversioned web" should still always check the latest stable release), but it is unclear where to grab that info from and as testing the info available when the workflow is triggered by a release requires doing a release, this is neigh impossible to test. I may need to create a dummy repo just to test.

For now, this at least fixes the Phive jobs.