GH Actions: auto-trigger update of schema website
Adds a workflow job to send out a GH API request whenever a new release is tagged.
The API request will trigger a workflow in the https://github.com/PHPCSStandards/schema.phpcodesniffer.com repository, which will update the schema.phpcodesniffer.com website.

As this is a cross-repo request, the standard GH secret is not sufficient.
A Personal Access Token has been created on the PHPCSStandards organisation to facilitate workflow dispatches. This PAT has been added to the repo actions secrets, so it can be safely used in this workflow.

The PAT token is a fine-grained one and needs the following permissions for this dispatch to succeed:
- contents: read & write
- metadata: read only (automatically selected when selecting the contents permission)

Note: The Personal Access Token needs to be renewed once a year.

Fixes 1094 (together with the newly set up https://github.com/PHPCSStandards/schema.phpcodesniffer.com repo