PSR2/PropertyDeclaration: update for properties with asymmetric visibility [2]
Per the FIG PSR-PER updates (next version), the "set-visibility MUST be _after_ the (read-)visibility".

This commit implements this, but as there may be two visibility modifiers, a minor refactor of the code involved in the wrong order checks for the `final` keyword and the `static` and `readonly` keywords is also needed as we can no longer presume that we can use the visibility keyword found with a search backwards to re-position other keywords.

Includes plenty of tests.
Includes a minor update to pre-existing test expectations (whitespace only), which is related to the refactor of the "fix"-code. As this sniff does not enforce any spacing requirements after the keywords, this difference is neglicable for the purpose of this sniff.

Note: in a future iteration, we may want to look at making the code for the "order" checks more efficient, but for now, at least it will work correctly again.

Refs:
* php-fig/per-coding-style 99
* php-fig/per-coding-style 114