Skip tests when 'git' command is not available
This change is intended to be a quick-fix to allow the test suite to pass. A
later enhancement would be to move this check to happen at runtime so that if a
user attempts to use one of these affected filters, a suitable error message
and exit code can be used. When that happens, the tests will need to be
reworked to cover that functionality.