Squiz/FunctionDeclarationArgumentSpacing: handle asym modifiers for constructor property promotion
This commit adds handling of the spacing after asymmetric visibility modifiers used for constructor property promotion to this sniff.

The spacing requirements are aligned with the spacing expectations of the `Squiz.WhiteSpace.ScopeKeywordSpacing` sniff, so the sniffs should not conflict with each other.

Additionally, the new check has a dedicated error code, which means that - if there would be a conflict anywhere - the asym visibility spacing check within this sniff can easily be turned off.

Includes tests.