fix: Check that the result of pcntl_waitpid is the PID of a managed process
The --parallel option spawns processes with pcntl_fork and expects the
result of pcntl_waitpid to be one of its spawned processes. It looks up
the temporary output file of a process in the given $childProcs array
without checking if the returned process ID is actually managed, which
under some circumstances caused by external factors may result in an
undefined array index error.

Fixes #1112