feat: Introduce bashunit test suite
This introduces the bashunit test suite as requested by @jrfnl. I opted to include the binary in the repository to make local testing easier & to lock the dependency on the project side instead of relying on external scripts.

This tests both a simple happy and a simple unhappy flow, as well as a test case for #1112.