Release checklist: improve information about regenerating the GPG key
Ran into some issues while trying to do this earlier today. Documenting my findings to prevent the same issues when having to do it again next year.
I've also asked a question about key rotation in the PHIVE repo to get clarification on some things: phar-io/phar.io 147. This may result in a new PHIVE documentation page about this being available by next year :crossed_fingers:

It is also the reason why I have not explicitly released the `3.13.1` tag yet as the README does not contain information about the updated GPG key yet.

By rights the new key should probably also be mentioned in the changelog, but tags should be immutable, so I'm going to leave that for now. Hopefully this change in the release checklist will prevent this snafu for next year.