Squiz/FunctionDeclarationArgumentSpacing: fix typo in `SpacingAfterSetVis[i]bility` error code
The `SpacingAfterVis[i]bility` error code was introduced in v 3.12.0 (March this year) for constructor property promotion.
The `SpacingAfterSetVis[i]bility` error code was introduced in v 3.13.1 (few days ago) for constructor property promotion with asym visibility.

Both contained the same typo.

Now, changing an error code is a breaking change. However, with 3.13.1 only having been released a few days ago, it is my opinion that if released soon as 3.13.2, this is an acceptable bug fix.
But then, it would be really weird to not fix the other error code in the same version. Then again, as PHPCS 4.0.0 is around the corner and does allow for breaking changes, let's leave the older typo for 4.0.0, just to be on the safe side.