Squiz/FunctionDeclarationArgumentSpacing: fix typo in `SpacingAfterVis[i]bility` error code
Follow up on #792, related to #1135

The `SpacingAfterVis[i]bility` error code was introduced in v 3.12.0 (March this year) for constructor property promotion.
The `SpacingAfterSetVis[i]bility` error code was introduced in v 3.13.1 (this month) for constructor property promotion with asym visibility.

Both contained the same typo.

PR 1135 fixes the asym error code and will be included in PHPCS 3.13.2.

This commit now fixes the non-asym error code and will be included in PHPCS 4.0.0.