Generic/InlineControlStructure: remove unreachable if condition
Commit e1b95951f377b5304f20ea03079876f9c9623c51 changed the
InlineControlStructure sniff to make it bail early for all control
structures without a body (before the sniff did that only for `while`
and `for`).

Before commit e1b95951f377b5304f20ea03079876f9c9623c51, `$nextNonEmpty`
could be `false` as there were some scenarios where there would be no
non-empty tokens after the `$stackPtr` or after the `parenthesis_closer`.
This is the case for the `else` without a body in
InlineControlStructureUnitTest.2.inc and the `if` without a body in
InlineControlStructureUnitTest.5.inc.

After commit e1b95951f377b5304f20ea03079876f9c9623c51, `$nextNonEmpty`
cannot be `false` anymore as cases where there are no non-empty tokens
are handled before this variable is set. This happens either in line 83
when `$nextTokenIndex` is set or line 98 when `$nextTokenIndex` is not
set (a control structure that is not `T_ELSE` or `T_DO` and is missing
the parenthesis closer).