PEAR/ObjectOperatorIndent: improve end of statement detection when getting the next object operator
This commit fixes a bug where `PEAR.WhiteSpace.ObjectOperatorIndent`
would generate a false positive when checking multiple chained object
operators in a multidimensional array.

The problem is that the code was using the `$local` parameter of
`findNext()` to limit the scope of the search when reassigning
`$next` to the next object operator, and `$local` only considers a
semicolon as the end of a statement.

Different chained calls are not necessarily separated by a semicolon (as
shown in the test added in this commit). This means that the sniff
currently considers the object operators on the second and third chained
method calls as part of the first chained method call. The sniff checks
their indentation using the indentation of the first element on the
first chained call as its base.

To fix this problem, instead of relying on the `$local` parameter to
find the end of the statement, the end of the statement as defined by
`File::findEndOfStatement()` is now used.