Generic/Syntax: add support for inspecting code passed via STDIN (#1151)
* Generic/Syntax: add support for inspecting code passed via STDIN

This commit improves the Generic.PHP.Syntax sniff to make it work when
code is passed via STDIN. Before, any code passed this way would cause a
false negative as the sniff was not taking into account that
`$phpcsFile->getFilename()` might return `STDIN` instead of an actual
file name.

The test is not run on Windows as STDIN is currently not supported on Windows.