PSR2/SwitchDeclaration: make "wrong opener" fixable if opener is a semi-colon
The PHP 8.5 deprecations RFC contains a proposal to deprecate the use of semicolons as scope openers for `switch` `case` and `default` statements.

The `PSR2.ControlStructures.SwitchDeclaration` sniff already detects and flags this issue under the `WrongOpenercase` and `WrongOpenerdefault` error codes, but as the same error code is also used for `case`/`default` statements using the `: { ... }` syntax, the error code was non-fixable.

With an eye on the PHP 8.5 deprecation, it seems prudent to ensure that a semi-colon `case`/`default` scope opener can be auto-fixed to a `colon`.

Making the `: { ... }` syntax auto-fixable may be more complex, so is left as future scope.

Includes tests.

Note: if this were a new sniff, I would probably have introduced new error code(s), so the fixable and the non-fixable errors would have separate error codes. However, as this is a pre-existing sniff, doing so would be a breaking change and not one I deem strictly necessary.

Ref:
* https://wiki.php.net/rfc/deprecations_php_8_5#deprecate_semicolon_after_case_in_switch_statement