GH Actions/verify release: auto-trigger Wiki update on new release
As of now, the Wiki is publicly editable via PRs to the PHPCSStandards/PHP_CodeSniffer-documentation repository.

The editable version of the Wiki includes some automation, like automatically updating various output examples displayed in the Wiki.

The resulting output may change between PHPCS versions, so this commit adds a job to the "verify release" workflow to automatically trigger a rebuild of the Wiki for each new PHPCS release.