AbstractMethodTestCase/AbstractTokenizerTestCase: use LocalFile instead of DummyFile
Should work just the same and just as well.

Historical context: looks like the use of `DummyFile` is a left over from a time long, long ago, when the "test case code" was included in the same file as the unit test class. This was changed in 2016 for PHPCS 2.x via PR squizlabs/PHP_CodeSniffer 1193, but then the merge up into PHPCS 3.x (which introduced `DummyFile`) did not change the object instantiation.